Move LocalFeedItem to it's own file
