+++ EssentialDeveloper/EssentialDeveloper/Feed Cache/FeedStore.swift
@@ -9,22 +9,3 @@ public protocol FeedStore {
     func insert(_ items: [LocalFeedItem], timestamp: Date, completion: @escaping InsertionCompletion)
 }
 
-public struct LocalFeedItem: Equatable {
-    public let id: UUID
-    public let description: String?
-    public let location: String?
-    public let imageURL: URL
-    
-    public init(
-        id: UUID,
-        description: String?,
-        location: String?,
-        imageURL: URL
-    ) {
-        self.id = id
-        self.description = description
-        self.location = location
-        self.imageURL = imageURL
-    }
-}
-
